backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
python-dotenv
torch